mnist_keras/requirements.txt
tensorflow
metaflow-checkpoint